requirements.txt
fastapi
uvicorn
pydantic-settings
openai
langchain
python-dotenv
ruff
pillow
numpy
opencv-python
matplotlib
mediapipe